+++ packages/core/src/adapters/monnify.ts
@@ -60,4 +60,7 @@ export const MonnifyAdapter: AdapterInterface = {
       },
     });
   },
+  getInstance: () => {
+    return window.MonnifySDK;
+  },
 };